add controller and models of servey
